+++ packages/graphql-yoga/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "@graphql-yoga/core",
-  "version": "0.1.0",
+  "name": "graphql-yoga",
+  "version": "2.0.0",
   "description": "",
   "repository": {
     "type": "git",
     "url": "https://github.com/dotansimha/graphql-yoga.git",
-    "directory": "packages/core"
+    "directory": "packages/graphql-yoga"
   },
   "main": "dist/index.js",
   "module": "dist/index.mjs",
@@ -43,15 +43,16 @@
   },
   "dependencies": {
     "@envelop/core": "^1.3.0",
-    "@envelop/disable-introspection": "^1.0.0",
-    "@envelop/parser-cache": "^2.0.0",
-    "@envelop/validation-cache": "^2.0.0",
+    "@graphql-yoga/core": "workspace:packages/core",
     "@graphql-yoga/handler": "workspace:packages/handler",
-    "ts-log": "^2.2.3"
+    "fastify": "^3.22.1",
+    "graphql": "^15.6.1",
+    "graphql-helix": "^1.8.3",
+    "pino": "^7.0.5",
+    "pino-pretty": "^7.1.0"
   },
   "devDependencies": {
     "@types/node": "^16.11.1",
-    "bob-the-bundler": "^1.5.1",
-    "graphql": "^15.6.1"
+    "bob-the-bundler": "^1.5.1"
   }
 }
