+++ packages/handler/package.json
@@ -15,9 +15,10 @@
   "author": "Saihajpreet Singh <saihajpreet.singh@gmail.com> (https://github.com/saihaj)",
   "license": "MIT",
   "devDependencies": {
+    "@envelop/core": "^1.3.0",
     "@types/node": "^16.11.1",
     "bob-the-bundler": "^1.5.1",
-    "fastify": "^3.22.1",
+    "graphql": "^15.6.1",
     "typescript": "^4.4.4"
   },
   "buildOptions": {
@@ -42,8 +43,6 @@
     "access": "public"
   },
   "dependencies": {
-    "@envelop/core": "^1.3.0",
-    "graphql": "^15.6.1",
     "graphql-helix": "^1.8.3"
   }
 }
